Make FILL_ME_IN dynamic, so casts are no lnger needed

+++ Koans/AboutStrings.cs
@@ -61,7 +61,7 @@ namespace DotNetCoreKoans.Koans
 am a
 broken line";
             var literalString = FILL_ME_IN;
-            Assert.Equal((int)FILL_ME_IN, verbatimString.Length);
+            Assert.Equal(FILL_ME_IN, verbatimString.Length);
             Assert.Equal(literalString, verbatimString);
         }
